feat : 뷰와 뷰컨트롤러 분리
